QA-6864 changes related to HQ Smoketest testcase 78
